fix: update test file according to dokojs version
